final HW1 Arithmetic Calculator
It works 100% autolab. First project of object-oriented programming.
Description: Its a java program that calculates every mathematic expression that contains operators and parenthesis. An example of a valid mathematic expression is: 5 + (((3.3 + 6.6) * 9.2 ) + 12.546) * 2.323 +
( ( ( 33.3 + 2342.1 ) * 55.555 ) - 10000.009 ) + 11.334 * 2.3 ^3.5
After that the program creates a binary tree that contains the operators and in the leaves the values and evaluate the result. In main you can type "-d" to print code that can be a graph using graphviz, and with "-c" you get the result. Also, using "-s" you get the postorder of the tree.